support for any type

+++ StaticChecker/Evaluator.cs
@@ -26,13 +26,13 @@ sealed class Evaluator
 
         if (node is BoundUnaryExpression u)
         {
-            var operand = (int)EvaluateExpression(u.Operand);
+            var operand = EvaluateExpression(u.Operand);
             switch (u.OperatorKind)
             {
                 case BoundUnaryOperatorKind.Identity:
-                    return operand;
+                    return (int)operand;
                 case BoundUnaryOperatorKind.Negation:
-                    return -operand;
+                    return -(int)operand;
                 default:
                     throw new Exception($"Unexpected union operator {u.OperatorKind}");
             }
@@ -40,19 +40,19 @@ sealed class Evaluator
 
         if (node is BoundBinaryExpression b)
         {
-            var left = (int)EvaluateExpression(b.Left);
-            var right = (int)EvaluateExpression(b.Right);
+            var left = EvaluateExpression(b.Left);
+            var right = EvaluateExpression(b.Right);
 
             switch (b.OperatorKind)
             {
                 case BoundBinaryOperatorKind.Addition:
-                    return left + right;
+                    return (int)left + (int)right;
                 case BoundBinaryOperatorKind.Subtraction:
-                    return left - right;
+                    return (int)left - (int)right;
                 case BoundBinaryOperatorKind.Multiplication:
-                    return left * right;
+                    return (int)left * (int)right;
                 case BoundBinaryOperatorKind.Division:
-                    return left / right;
+                    return (int)left / (int)right;
                 default:
                     throw new Exception($"Unexpected binary operator {b.OperatorKind}");
             }
